epdk-platform/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
structlog==24.1.0
prometheus-fastapi-instrumentator==7.1.0
python-jose[cryptography]==3.3.0
slowapi==0.1.9
httpx==0.27.0
